refactor, simplify
